# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Death Dance API Documentation",
        "description": "Death Dance doc of all endpoints",
        "contact": {
            "email": "luciano.meniconi.r@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get list of appointments",
                "description": "Returns list of appointments",
                "operationId": "getAppointmentsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Store new appointment",
                "description": "Returns appointment data",
                "operationId": "storeAppointment",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment information",
                "description": "Returns appointment data",
                "operationId": "getAppointmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update existing appointment",
                "description": "Returns updated appointment data",
                "operationId": "updateAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete existing appointment",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/appointments/{date}/hours": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get list of available hours",
                "description": "Returns list of available hours in a  date",
                "operationId": "getHoursList",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Appointments",
            "description": "API Endpoints of appointments"
        }
    ],
    "security": [
        []
    ]
}